fix: `terraform_version` not being handled by `terraform/install` correctly
The `terraform_version` parameter for the `install` command is supposed to
support fuzzy matching of the version, however in practice this functionality
is broken because the raw version string is passed to almost all of the
functions. Passing the resolved version (from `determine_version`) into those
functions fixes the issue.

Fixes #87